+++ data/repairshop.trace.db
@@ -81,8 +81,8 @@ call next value for hibernate_sequence [90036-214]
 	at org.springframework.aop.framework.ReflectiveMethodInvocation.proceed(ReflectiveMethodInvocation.java:186)
 	at org.springframework.aop.framework.JdkDynamicAopProxy.invoke(JdkDynamicAopProxy.java:215)
 	at com.sun.proxy.$Proxy107.save(Unknown Source)
-	at com.pmm.budgetcalculator.expensing.domain.expense.ExpenseService.createExpense(ExpenseService.java:23)
-	at com.pmm.budgetcalculator.expensing.application.ExpenseController.createExpense(ExpenseController.java:29)
+	at ExpenseService.createExpense(ExpenseService.java:23)
+	at com.pmm.budgetcalculator.expensing.controllers.ExpenseController.createExpense(ExpenseController.java:29)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
 	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
@@ -216,8 +216,8 @@ call next value for hibernate_sequence [90036-214]
 	at org.springframework.aop.framework.ReflectiveMethodInvocation.proceed(ReflectiveMethodInvocation.java:186)
 	at org.springframework.aop.framework.JdkDynamicAopProxy.invoke(JdkDynamicAopProxy.java:215)
 	at com.sun.proxy.$Proxy107.save(Unknown Source)
-	at com.pmm.budgetcalculator.expensing.domain.expense.ExpenseService.createExpense(ExpenseService.java:23)
-	at com.pmm.budgetcalculator.expensing.application.ExpenseController.createExpense(ExpenseController.java:29)
+	at ExpenseService.createExpense(ExpenseService.java:23)
+	at com.pmm.budgetcalculator.expensing.controllers.ExpenseController.createExpense(ExpenseController.java:29)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
 	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
@@ -351,8 +351,8 @@ call next value for hibernate_sequence [90036-214]
 	at org.springframework.aop.framework.ReflectiveMethodInvocation.proceed(ReflectiveMethodInvocation.java:186)
 	at org.springframework.aop.framework.JdkDynamicAopProxy.invoke(JdkDynamicAopProxy.java:215)
 	at com.sun.proxy.$Proxy107.save(Unknown Source)
-	at com.pmm.budgetcalculator.expensing.domain.expense.ExpenseService.createExpense(ExpenseService.java:23)
-	at com.pmm.budgetcalculator.expensing.application.ExpenseController.createExpense(ExpenseController.java:29)
+	at ExpenseService.createExpense(ExpenseService.java:23)
+	at com.pmm.budgetcalculator.expensing.controllers.ExpenseController.createExpense(ExpenseController.java:29)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
 	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
