--- src/main/java/com/pmm/budgetcalculator/expensing/domain/expense/ExpenseFactory.java
@@ -1,13 +1,0 @@
-package com.pmm.budgetcalculator.expensing.domain.expense;
-
-import com.pmm.budgetcalculator.foundation.AbstractFactory;
-import org.springframework.stereotype.Component;
-
-@Component
-public class ExpenseFactory implements AbstractFactory<Expense> {
-
-    @Override
-    public Expense create() {
-        return new Expense();
-    }
-}
