+++ src/main/java/com/pmm/budgetcalculator/expensing/services/ExpenseService.java
@@ -1,5 +1,7 @@
-package com.pmm.budgetcalculator.expensing.domain.expense;
+package com.pmm.budgetcalculator.expensing.services;
 
+import com.pmm.budgetcalculator.expensing.repositories.Expense;
+import com.pmm.budgetcalculator.expensing.repositories.ExpenseRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
