--- src/main/java/com/pmm/budgetcalculator/foundation/AbstractFactory.java
@@ -1,7 +1,0 @@
-package com.pmm.budgetcalculator.foundation;
-
-public interface AbstractFactory<E extends BaseEntity> {
-
-  E create();
-
-}
