--- src/main/java/com/pmm/budgetcalculator/foundation/AbstractService.java
@@ -1,8 +1,0 @@
-package com.pmm.budgetcalculator.foundation;
-
-public interface AbstractService<E extends BaseEntity> {
-
-  E add(E entity);
-  E change(E entity);
-
-}
